refactor(mirror-schema): rename app-key.ts to api-key.ts
