+++ mirror/mirror-server/src/functions/keys/create.function.ts
@@ -12,7 +12,7 @@ import {
   isValidAppKeyName,
   normalizePermissions,
   type Permissions,
-} from 'mirror-schema/src/app-key.js';
+} from 'mirror-schema/src/api-key.js';
 import {randomBytes} from 'node:crypto';
 import {appAuthorization, userAuthorization} from '../validators/auth.js';
 import {validateSchema} from '../validators/schema.js';
