+++ mirror/mirror-server/src/functions/keys/delete.function.ts
@@ -3,7 +3,7 @@ import {
   deleteAppKeysRequestSchema,
   deleteAppKeysResponseSchema,
 } from 'mirror-protocol/src/app-keys.js';
-import {appKeyPath} from 'mirror-schema/src/app-key.js';
+import {appKeyPath} from 'mirror-schema/src/api-key.js';
 import {appAuthorization, userAuthorization} from '../validators/auth.js';
 import {validateSchema} from '../validators/schema.js';
 import {userAgentVersion} from '../validators/version.js';
