+++ mirror/mirror-server/src/functions/keys/list.function.ts
@@ -7,7 +7,7 @@ import {
   ALL_PERMISSIONS,
   appKeyDataConverter,
   appKeysCollection,
-} from 'mirror-schema/src/app-key.js';
+} from 'mirror-schema/src/api-key.js';
 import {appAuthorization, userAuthorization} from '../validators/auth.js';
 import {validateSchema} from '../validators/schema.js';
 import {userAgentVersion} from '../validators/version.js';
