+++ mirror/mirror-server/src/functions/validators/https.test.ts
@@ -10,7 +10,7 @@ import {
   Permissions,
   appKeyDataConverter,
   appKeyPath,
-} from 'mirror-schema/src/app-key.js';
+} from 'mirror-schema/src/api-key.js';
 import {appPath} from 'mirror-schema/src/deployment.js';
 import {setApp, setUser} from 'mirror-schema/src/test-helpers.js';
 import {userPath} from 'mirror-schema/src/user.js';
